Restructuring of ReferenceContext and ReadWalkers to accept a ReferenceContext.  Now ReferenceContext is byte[] backed not char[].  Please no more chars for the reference.  All of the tests pass now.  Coming check-ins are going to clean up the char / byte problems in the GATK
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@3397 348d0f76-0448-11de-a6fe-93d51630548a

+++ java/src/org/broadinstitute/sting/oneoffprojects/walkers/AlleleBalanceHistogramWalker.java
@@ -104,7 +104,7 @@ public class AlleleBalanceHistogramWalker extends LocusWalker<Map<String,Double>
         }
         
         for ( PileupElement e : alicon.getBasePileup() ) {
-            if ( BaseUtils.basesAreEqual( e.getBase(), (byte) ref.getBase() ) ) {
+            if ( BaseUtils.basesAreEqual( e.getBase(), ref.getBase() ) ) {
                 refBases++;
             } else if ( BaseUtils.basesAreEqual(e.getBase(), (byte) snpBase ) ) {
                 altBases++;
